// src/app/features/transactiontable/tblPurchaseOrder/models/tblPurchaseOrder-Add.model.ts
export interface TblPurchaseOrderAdd {
	fldId: number;
	fldFKUnitId: number;
	fldPONo: string;
	fldPODate: Date;
	fldFKSupplierID: number;
	fldDeliveryStartDate: Date;
	fldDeliveryEndDate: Date;
	fldTotalQty: number;
	fldInwardQty: number;
	fldCancelQty: number;
	fldBalanceQty: number;
	fldRemarks: string;
	fldFKStatus: number;
	fldItemsGrossValue: number;
	fldItemsTaxableValue: number;
	fldItemsGSTValue: number;
	fldItemsTotalValue: number;
	fldDiscountPercentage: number;
	fldDiscountValue: number;
	fldGrandGrossValue: number;
	fldOtherPlusValue: number;
	fldOtherMinusValue: number;
	fldRoundoff: number;
	fldGrandTotalValue: number;
	fldPaidAmount: number;
	fldFKPaymentStatus: number;
	fldIsActive: boolean;
	fldCreatedBy: number;
	fldCreatedDt: Date;
}